+++ order/src/main/java/com/example/order/entity/response/OrderResponse.java
@@ -2,5 +2,7 @@ package com.example.order.entity.response;
 
 import com.example.common.entity.response.UserResponse;
 
-public record OrderResponse(long id, int count, UserResponse userResponse) {
+import java.math.BigInteger;
+
+public record OrderResponse(BigInteger id, int count, UserResponse userResponse) {
 }
