+++ product/src/main/java/com/example/product/entity/response/ProductResponse.java
@@ -1,3 +1,5 @@
 package com.example.product.entity.response;
 
-public record ProductResponse(Long id, String name, Integer status, Integer dirty) {}
+import java.math.BigInteger;
+
+public record ProductResponse(BigInteger id, String name, Integer status, Integer dirty) {}
